Move more helpers to util
